requirements.txt
alembic==1.13.1
bcrypt==4.1.2
bleach==6.1.0
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-FlatPages==0.8.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==22.0.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
Markdown==3.9
MarkupSafe==2.1.5
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.1
SQLAlchemy==2.0.27
typing_extensions==4.9.0
waitress==3.0.0
Werkzeug==3.0.1
WTForms==3.1.2
zipp==3.17.0
